Add first cut of the client.js and ts defs
NOTE: this is not a general purpose client library, just meant to be used on the web page.  Can't really use this outside of the webpage.  Rewrite to handle the websocket connection, and take functions for callbacks, or use an event emitter.  We just need to handle all of the websocket states